+++ nodes/tibbermodule.js
@@ -60,7 +60,7 @@ module.exports = {
     try {
       let response = await fetch(url, payload);
       if (response.status < 200 || response.status >= 300) {
-        return { 'error': true, 'details': { 'status': response.status, 'statusText': response.statusText } };
+        return { 'error': true, 'details': { 'status': response.status, 'statusText': response.statusText, 'response': response } };
       }
       let json = await response.text();
       // console.dir(json);
@@ -76,6 +76,10 @@ module.exports = {
       return { 'error': true, 'details': error };
     }
   },
+  // Use WS to subscribe to a stream for some data - future release
+  getSubscription: function (url, payload) {
+
+  },
   // Function for getting users home data
   get: function (query) {
     return this.getData(url, this.createRequest(queries[query]));
@@ -83,5 +87,9 @@ module.exports = {
   // function for sending your own query
   query: function (query) {
     return this.getData(url, this.createRequest(query));
+  },
+  // function for getting a subscription/stream instead (not implemented yet)
+  subscribe: function (query) {
+    return this.getSubscription(url, this.createRequest(query));
   }
 };